Turned more ints into floats to fix damage reduction

+++ OlympusUnity/Assets/_Project/Scripts/Characters/Gods/GodBehaviour.cs
@@ -112,7 +112,7 @@ public class GodBehaviour : MonoBehaviour
         }
     }
 
-    public virtual void OnDamageEvent(int damageTaken)
+    public virtual void OnDamageEvent(float damageTaken)
     {
         // Override if needed
     }
